dbm_api/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
PyMySQL==0.7.11
python-decouple==3.1
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.1.14
uritemplate==3.0.0
urllib3==1.22
